Teach build scripts to create frameworks
Frameworks are useful to help iOS developers to easily integrate
our stuff, therefore here it is a commit teaching our build scripts
to generate frameworks containing all you need to use our code.

This is still a rather low level repository hence the output is a
number of frameworks. I've gone down this path (rather than trying
to make nested frameworks) since I've read it's not advisable to
make nested frameworks with static libraries.

I've tested that this works by referencing to the frameworks from
a small project that I use for testing MeasurementKit on iOS.